[BE] refactor : 리뷰 반영

+++ backend/reviewduck/src/main/java/com/reviewduck/service/ReviewService.java
@@ -13,7 +13,7 @@ import com.reviewduck.domain.QuestionAnswer;
 import com.reviewduck.domain.Review;
 import com.reviewduck.domain.ReviewForm;
 import com.reviewduck.dto.request.AnswerRequest;
-import com.reviewduck.dto.request.ReviewCreateRequest;
+import com.reviewduck.dto.request.ReviewRequest;
 import com.reviewduck.exception.NotFoundException;
 import com.reviewduck.repository.QuestionRepository;
 import com.reviewduck.repository.ReviewRepository;
@@ -33,7 +33,7 @@ public class ReviewService {
         this.questionRepository = questionRepository;
     }
 
-    public Review save(String code, ReviewCreateRequest request) {
+    public Review save(String code, ReviewRequest request) {
         ReviewForm reviewForm = reviewFormService.findByCode(code);
 
         List<QuestionAnswer> questionAnswers = convertToQuestionAnswers(request.getAnswers());
@@ -58,7 +58,7 @@ public class ReviewService {
         return reviewRepository.findByReviewForm(reviewForm);
     }
 
-    public Review update(Long id, ReviewCreateRequest request) {
+    public Review update(Long id, ReviewRequest request) {
         Review review = reviewRepository.findById(id)
             .orElseThrow(() -> new NotFoundException("존재하지 않는 회고입니다."));
 
@@ -67,6 +67,9 @@ public class ReviewService {
     }
 
     public void delete(Long id) {
+        if (!reviewRepository.existsById(id)) {
+            throw new NotFoundException("존재하지 않는 회고입니다.");
+        }
         reviewRepository.deleteById(id);
     }
 }
